fix(test): serialize integration server port allocation across binaries
