+++ ailoop-cli/Cargo.toml
@@ -33,3 +33,4 @@ async-trait = { workspace = true }
 [dev-dependencies]
 reqwest = { version = "0.11", default-features = false, features = ["json", "rustls-tls"] }
 tempfile = { workspace = true }
+fs2 = { workspace = true }
